restructure `tools.vitruv.change.atomic` to expose as few as possible
