+++ bundles/tools.vitruv.change.atomic/src/tools/vitruv/change/atomic/command/internal/ApplyForwardCommandSwitch.xtend
@@ -1,4 +1,4 @@
-package tools.vitruv.change.atomic.command
+package tools.vitruv.change.atomic.command.internal
 
 import edu.kit.ipd.sdq.activextendannotations.Utility
 import java.util.List
@@ -8,6 +8,7 @@ import org.eclipse.emf.ecore.EObject
 import org.eclipse.emf.edit.command.AddCommand
 import org.eclipse.emf.edit.command.SetCommand
 import tools.vitruv.change.atomic.EChange
+import tools.vitruv.change.atomic.command.RemoveAtCommand
 import tools.vitruv.change.atomic.eobject.CreateEObject
 import tools.vitruv.change.atomic.eobject.DeleteEObject
 import tools.vitruv.change.atomic.feature.UnsetFeature
@@ -20,8 +21,8 @@ import tools.vitruv.change.atomic.feature.reference.ReplaceSingleValuedEReferenc
 import tools.vitruv.change.atomic.root.InsertRootEObject
 import tools.vitruv.change.atomic.root.RemoveRootEObject
 
-import static extension tools.vitruv.change.atomic.command.ChangeCommandUtil.alreadyContainsObject
-import static extension tools.vitruv.change.atomic.command.ChangeCommandUtil.getEditingDomain
+import static extension tools.vitruv.change.atomic.command.internal.ChangeCommandUtil.alreadyContainsObject
+import static extension tools.vitruv.change.atomic.command.internal.ChangeCommandUtil.getEditingDomain
 
 /**
  * Switch to create commands for all EChange classes.
