+++ bundles/tools.vitruv.change.interaction.model/META-INF/MANIFEST.MF
@@ -9,6 +9,7 @@ Bundle-Vendor: %providerName
 Bundle-Localization: plugin
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Export-Package: tools.vitruv.change.interaction,
+ tools.vitruv.change.interaction.impl,
  tools.vitruv.change.interaction.util
 Require-Bundle: org.eclipse.emf.ecore;visibility:=reexport,
  org.eclipse.core.runtime
