+++ tests/tools.vitruv.change.atomic.tests/src/tools/vitruv/change/atomic/command/internal/CommandTest.xtend
@@ -1,4 +1,4 @@
-package tools.vitruv.change.atomic.command
+package tools.vitruv.change.atomic.command.internal
 
 import org.eclipse.emf.common.command.Command
 import static org.junit.jupiter.api.Assertions.assertTrue
